# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Speedtest Tracker API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/results/latest": {
            "get": {
                "tags": [
                    "Results"
                ],
                "summary": "Fetch the single most recent result",
                "description": "Requires an API token with scope `results:read`.",
                "operationId": "getLatestResult",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/results": {
            "get": {
                "tags": [
                    "Results"
                ],
                "summary": "List results",
                "description": "Requires an API token with scope `results:read`.",
                "operationId": "listResults",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultsCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ookla/list-servers": {
            "get": {
                "tags": [
                    "Servers"
                ],
                "summary": "List available Ookla speedtest servers",
                "description": "Requires an API token with scope `ookla:list-servers`.",
                "operationId": "listSpeedtestServers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServersCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                },
                                "example": {
                                    "message": "You do not have permission to view speedtest servers."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/speedtests/run": {
            "post": {
                "tags": [
                    "Speedtests"
                ],
                "summary": "Run a new Ookla speedtest",
                "description": "Requires an API token with scope `speedtests:run`.",
                "operationId": "runSpeedtest",
                "parameters": [
                    {
                        "name": "server_id",
                        "in": "query",
                        "description": "Optional Ookla speedtest server ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QueuedResultResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                },
                                "example": {
                                    "message": "You do not have permission to run speedtests."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/results/{id}": {
            "get": {
                "tags": [
                    "Results"
                ],
                "summary": "Fetch a single result by ID",
                "description": "Requires an API token with scope `results:read`.",
                "operationId": "getResult",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the result",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Result not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/stats": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Fetch aggregated Speedtest statistics",
                "description": "Requires an API token with scope `results:read`.",
                "operationId": "getStats",
                "parameters": [
                    {
                        "name": "start_at",
                        "in": "query",
                        "description": "ISO‑8601 start datetime filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "end_at",
                        "in": "query",
                        "description": "ISO‑8601 end datetime filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Stats"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ForbiddenError": {
                "description": "Forbidden error response when user lacks permission",
                "properties": {
                    "message": {
                        "description": "Error message indicating lack of permission",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "NotFoundError": {
                "description": "Error when a requested result is not found",
                "properties": {
                    "message": {
                        "description": "Result not found error message",
                        "type": "string",
                        "example": "Result not found."
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "description": "Pagination navigation links",
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "description": "Pagination metadata",
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "QueuedResultData": {
                "description": "Additional data for a queued result",
                "properties": {
                    "server": {
                        "$ref": "#/components/schemas/QueuedResultDataServer"
                    }
                },
                "type": "object"
            },
            "QueuedResultDataServer": {
                "description": "Server info inside queued result data",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "QueuedResultResponse": {
                "description": "Response when a speedtest is queued",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/QueuedResult"
                    },
                    "message": {
                        "description": "Response status message",
                        "type": "string",
                        "example": "Speedtest added to the queue."
                    }
                },
                "type": "object"
            },
            "QueuedResult": {
                "description": "A queued speedtest result",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "service": {
                        "type": "string"
                    },
                    "ping": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "download": {
                        "type": "integer",
                        "nullable": true
                    },
                    "upload": {
                        "type": "integer",
                        "nullable": true
                    },
                    "benchmarks": {
                        "type": "object",
                        "nullable": true
                    },
                    "healthy": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "status": {
                        "type": "string"
                    },
                    "scheduled": {
                        "type": "boolean"
                    },
                    "comments": {
                        "type": "string",
                        "nullable": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/QueuedResultData"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ResultResponse": {
                "description": "Envelope for a single result response",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Result"
                    },
                    "message": {
                        "description": "Response status message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Result": {
                "description": "A single Speedtest result entry",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 3
                    },
                    "service": {
                        "type": "string",
                        "example": "ookla"
                    },
                    "ping": {
                        "type": "number",
                        "format": "float",
                        "example": 6.246,
                        "nullable": true
                    },
                    "download": {
                        "type": "integer",
                        "example": 11784522
                    },
                    "upload": {
                        "type": "integer",
                        "example": 3596514
                    },
                    "download_bits": {
                        "type": "integer",
                        "example": 94276176
                    },
                    "upload_bits": {
                        "type": "integer",
                        "example": 28772112
                    },
                    "download_bits_human": {
                        "type": "string",
                        "example": "94.28 Mbps"
                    },
                    "upload_bits_human": {
                        "type": "string",
                        "example": "28.77 Mbps"
                    },
                    "benchmarks": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "healthy": {
                        "type": "boolean",
                        "example": null,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "completed"
                    },
                    "scheduled": {
                        "type": "boolean",
                        "example": false
                    },
                    "comments": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "data": {
                        "description": "Raw speedtest details",
                        "type": "object",
                        "example": {
                            "isp": "Speedtest Communications",
                            "ping": {
                                "low": 17.841,
                                "high": 24.077,
                                "jitter": 1.878,
                                "latency": 19.133
                            },
                            "type": "result",
                            "result": {
                                "id": "abc123-uuid",
                                "url": "https://docs.speedtest-tracker.dev",
                                "persisted": true
                            },
                            "server": {
                                "id": 1234,
                                "ip": "127.0.0.1",
                                "host": "docs.speedtest-tracker.dev",
                                "name": "Speedtest",
                                "port": 8080,
                                "country": "United States",
                                "location": "New York City, NY"
                            },
                            "upload": {
                                "bytes": 124297377,
                                "elapsed": 9628,
                                "latency": {
                                    "iqm": 341.111,
                                    "low": 16.663,
                                    "high": 529.86,
                                    "jitter": 37.587
                                },
                                "bandwidth": 113750000
                            },
                            "download": {
                                "bytes": 230789788,
                                "elapsed": 14301,
                                "latency": {
                                    "iqm": 104.125,
                                    "low": 23.72,
                                    "high": 269.563,
                                    "jitter": 13.447
                                },
                                "bandwidth": 115625000
                            },
                            "interface": {
                                "name": "eth0",
                                "isVpn": false,
                                "macAddr": "00:00:00:00:00:00",
                                "externalIp": "127.0.0.1",
                                "internalIp": "127.0.0.1"
                            },
                            "timestamp": "2025-05-21T12:00:00Z",
                            "packetLoss": 11
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-21 17:44:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-21 17:44:16"
                    }
                },
                "type": "object"
            },
            "ResultsCollection": {
                "description": "Paginated list of Results",
                "properties": {
                    "data": {
                        "description": "Array of result objects",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Result"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    }
                },
                "type": "object"
            },
            "ServersCollection": {
                "description": "Mapping of server IDs to display names",
                "properties": {
                    "data": {
                        "description": "Map of server ID to server display name",
                        "type": "object",
                        "example": {
                            "20001": "Washington Fiber (Washington, DC, 20001)",
                            "90001": "LA Connect (Los Angeles, CA, 90001)",
                            "60601": "ChiTown Broadband (Chicago, IL, 60601)",
                            "77001": "Houston Net (Houston, TX, 77001)",
                            "33101": "MiamiLink (Miami, FL, 33101)"
                        },
                        "additionalProperties": true
                    },
                    "message": {
                        "description": "Response status message",
                        "type": "string",
                        "example": "Speedtest servers fetched successfully."
                    }
                },
                "type": "object"
            },
            "Stats": {
                "description": "Aggregated speedtest statistics",
                "properties": {
                    "total_results": {
                        "type": "integer"
                    },
                    "avg_ping": {
                        "type": "number",
                        "format": "float"
                    },
                    "avg_download": {
                        "type": "number",
                        "format": "float"
                    },
                    "avg_upload": {
                        "type": "number",
                        "format": "float"
                    },
                    "min_ping": {
                        "type": "number",
                        "format": "float"
                    },
                    "min_download": {
                        "type": "number",
                        "format": "float"
                    },
                    "min_upload": {
                        "type": "number",
                        "format": "float"
                    },
                    "max_ping": {
                        "type": "number",
                        "format": "float"
                    },
                    "max_download": {
                        "type": "number",
                        "format": "float"
                    },
                    "max_upload": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "UnauthenticatedError": {
                "description": "Error when user is not authenticated",
                "properties": {
                    "message": {
                        "description": "Unauthenticated error message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "description": "Validation failed due to invalid server_id input",
                "properties": {
                    "message": {
                        "description": "Validation failed due to invalid server_id input",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "Accept": {
                "name": "Accept",
                "in": "header",
                "description": "Expected response format",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "application/json"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Results",
            "description": "Operations related to Speedtest results"
        },
        {
            "name": "Speedtests",
            "description": "Operations to run speedtests"
        },
        {
            "name": "Servers",
            "description": "Operations for speedtest servers"
        },
        {
            "name": "Stats",
            "description": "Operations for statistics"
        }
    ]
}